+++ backend/orchestrator/computetask.py
@@ -1,17 +1,17 @@
 import copy
 
 
-def orc_to_localrep(data: dict) -> dict:
-    """Convert a compute task from the orchestrator format to the localrep format"""
+def orc_to_api(data: dict) -> dict:
+    """Convert a compute task from the orchestrator format to the api format"""
 
     res = copy.deepcopy(data)
-    res["inputs"] = [_input_to_localrep(input) for input in res["inputs"]]
+    res["inputs"] = [_input_to_api(input) for input in res["inputs"]]
     res["outputs"] = [{"identifier": identifier, **output} for identifier, output in res["outputs"].items()]
     return res
 
 
-def _input_to_localrep(input: dict) -> dict:
-    """Convert a compute task input from the orchestrator format to the localrep format"""
+def _input_to_api(input: dict) -> dict:
+    """Convert a compute task input from the orchestrator format to the api format"""
 
     if "parent_task_output" not in input:
         return input