+++ backend/organization/views/organization.py
@@ -1,9 +1,9 @@
 from rest_framework import mixins
 from rest_framework.viewsets import ModelViewSet
 
-from localrep.models import ChannelOrganization as ChannelOrganizationRep
-from localrep.serializers import ChannelOrganizationSerializer as ChannelOrganizationRepSerializer
-from localrep.views.utils import get_channel_name
+from api.models import ChannelOrganization as ChannelOrganizationRep
+from api.serializers import ChannelOrganizationSerializer as ChannelOrganizationRepSerializer
+from api.views.utils import get_channel_name
 
 
 class OrganizationViewSet(ModelViewSet, mixins.ListModelMixin):