+++ backend/substrapp/compute_tasks/outputs.py
@@ -7,8 +7,8 @@ from django.urls import reverse
 
 import orchestrator
 import orchestrator.model_pb2 as model_pb2
-from localrep.errors import AlreadyExistsError
-from localrep.serializers import ModelSerializer as ModelRepSerializer
+from api.errors import AlreadyExistsError
+from api.serializers import ModelSerializer as ModelRepSerializer
 from substrapp.compute_tasks import context
 from substrapp.compute_tasks import directories
 from substrapp.compute_tasks.asset_buffer import add_model_from_path
@@ -57,28 +57,28 @@ class OutputSaver:
     def _save_models(self, models: list[context.OutputResource]):
         logger.info("saving models")
 
-        local_models = []
+        new_models = []
         for model in models:
-            local_model = self._save_model_to_local_storage(model)
-            local_models.append(local_model)
+            new_model = self._save_model_to_local_storage(model)
+            new_models.append(new_model)
 
         try:
             with get_orchestrator_client(self._ctx.channel_name) as client:
-                localrep_data_models = client.register_models({"models": local_models})
+                registered_models = client.register_models({"models": new_models})
         except Exception as exc:
-            for model in local_models:
+            for model in new_models:
                 self._delete_model(model["key"])
             raise exc
 
-        for localrep_data in localrep_data_models:
-            localrep_data["channel"] = self._ctx.channel_name
-            localrep_serializer = ModelRepSerializer(data=localrep_data)
+        for registered_model in registered_models:
+            registered_model["channel"] = self._ctx.channel_name
+            serializer = ModelRepSerializer(data=registered_model)
             try:
-                localrep_serializer.save_if_not_exists()
+                serializer.save_if_not_exists()
             except AlreadyExistsError:
                 pass
 
-        for model, local in zip(models, local_models):
+        for model, local in zip(models, new_models):
             path = os.path.join(self._ctx.directories.task_dir, model.rel_path)
             add_model_from_path(path, local["key"])
 
@@ -92,7 +92,7 @@ class OutputSaver:
         with open(path, "rb") as f:
             instance.file.save("model", f)
         current_site = settings.DEFAULT_DOMAIN
-        storage_address = f'{current_site}{reverse("localrep:model-file", args=[instance.key])}'
+        storage_address = f'{current_site}{reverse("api:model-file", args=[instance.key])}'
 
         logger.debug("Saving model in local storage", model_key=instance.key, identifier=model.identifier)